Manage plugins with nix again
My configuration is simply symlinked to the repository, so I can still
edit my config and tweak it without having to reinstall everything.

Treesitter breaking all the time has become annoying. It has also been
slightly annoying to update lazy in a separate step. The move to lazy
was motivated by the ability to actually run tests for plugins and work
on them. Since I rarely work on plugins, I'll just have to setup a
docker container to run a normal neovim version and test it.