Use rustaceanvim for neotest
neotest-rust doesn't work correctly with the versions in nixpkgs.

The version in nixpkgs for rustaceanvim is also not recent eneough to
have the neotest module and hence it is set as a flake.

One nice thing about this integration is that rstest's cases can now be
started with neotest. This might also be available with neotest-rust if
the versions were compatible, but that is something I can neither
confirm nor deny.

Since I'm using rustaceanvim anyway, this also means one less package to
install.