Remove telescope-project
The intent was to use it for neovide to quickly jump to projects. While
that works, the tooling for the project is not available becase I'm
provisioning it with flakes. I therefore have to start neovide from a
terminal that has the tooling initialized or figure out a different way
to achieve the same.

Revert 1 commits

8c588d6 'Add telescope-project'